Add Firebase Storage Plugin
This commit adds the Firebase Storage plugin to the project. The necessary changes include updating the `GeneratedPluginRegistrant.swift` for macOS, modifying `pubspec.lock` and `pubspec.yaml` to include the Firebase Storage dependency, and updating the `generated_plugin_registrant.cc` and `generated_plugins.cmake` for Windows to register the Firebase Storage plugin.